requirements.txt
sklearn
numpy
pandas
matplotlib
scipy
seaborn
keras
tensorflow-gpu
Pillow